fix: ImgServiceImpl 주석처리
